*: rename `neofs-lens` to `neofs-lancet`
Reflect the new CLI name across the codebase and release notes. The tool now
covers storage modifications in addition to inspection.

Closes #3822.
