+++ CHANGELOG.md
@@ -4,13 +4,13 @@ Changelog for NeoFS Node
 ## [Unreleased]
 
 ### Added
-- `neofs-lens meta resync` command (#3849, #3890)
+- `neofs-lancet meta resync` command (#3849, #3890)
 - `policer.boost_multiplier` SN config option (#3855)
-- `neofs-lens storage flush-write-caches` command (#3872)
+- `neofs-lancet storage flush-write-caches` command (#3872)
 - Reload gRPC SN config with SIGHUP (#3874)
 - Support for creation of containers with initial placement policy (#3830)
-- `neofs-lens meta remove` command (#3891)
-- `neofs-lens fstree remove` command (#3892)
+- `neofs-lancet meta remove` command (#3891)
+- `neofs-lancet fstree remove` command (#3892)
 
 ### Fixed
 - Resending the header after chunks have already been sent in object service `Get` handler (#3833)
@@ -27,7 +27,7 @@ Changelog for NeoFS Node
 
 ### Changed
 - SN returns unsigned responses to requests with API >= `v2.22` (#3785)
-- Move lens `write-cache` `get`/`list` commands into `fstree` section in CLI (#3838, #3844)
+- Move lancet `write-cache` `get`/`list` commands into `fstree` section in CLI (#3838, #3844)
 - Session key storage access session keys by account only (#3817)
 - Optimized locking for reputation data collector (#3851)
 - Optimized local HEAD/GET request execution (#3783, #3882)
@@ -37,6 +37,7 @@ Changelog for NeoFS Node
 - Policer starts from a random offset (#3879)
 - SN now determines request API version from original request meta header (#3897)
 - 2.18 is the minimum supported version for new objects (#3869)
+- Rename `neofs-lens` into `neofs-lancet` (#3894)
 
 ### Removed
 - `node.persistent_sessions.path` config option from SN config (#3846)
@@ -65,7 +66,7 @@ and has now been removed. Drop this config and associated file, migrate with 0.5
 Sessions are stored in `node.persistent_state.path` DB now.
 
 Delete `storage.shards.resync_metabase` config option from SN config, it's no longer used.
-Use `neofs-lens meta resync` command if you need to resync metabase.
+Use `neofs-lancet meta resync` command if you need to resync metabase.
 
 Storage nodes no longer automatically migrate metabases from version 5
 (NeoFS 0.46.0) to 6 (NeoFS 0.48.0) and from version 6 to version 7 (NeoFS
@@ -81,6 +82,10 @@ supporting NeoFS API version starting from `v2.21`, the flag may be safely
 dropped. For older servers, success checks should now be done by polling with
 symmetric read command instead. Please note that this flag will be removed in a
 future release.
+
+`neofs-lens` command was renamed to `neofs-lancet` to reflect that it now
+supports modifying storage state in addition to inspection.
+Replace `neofs-lens` references with `neofs-lancet`.
 
 ## [0.51.1] - 2026-02-18
