+++ cmd/neofs-lancet/internal/fstree/cleanup.go
@@ -1,7 +1,7 @@
 package fstree
 
 import (
-	common "github.com/nspcc-dev/neofs-node/cmd/neofs-lens/internal"
+	common "github.com/nspcc-dev/neofs-node/cmd/neofs-lancet/internal"
 	"github.com/spf13/cobra"
 )
