+++ cmd/neofs-lancet/internal/fstree/list.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"io"
 
-	common "github.com/nspcc-dev/neofs-node/cmd/neofs-lens/internal"
+	common "github.com/nspcc-dev/neofs-node/cmd/neofs-lancet/internal"
 	oid "github.com/nspcc-dev/neofs-sdk-go/object/id"
 	"github.com/spf13/cobra"
 )
