+++ cmd/neofs-lancet/internal/fstree/root.go
