+++ cmd/neofs-lancet/internal/meta/id.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 
 	"github.com/mr-tron/base58"
-	common "github.com/nspcc-dev/neofs-node/cmd/neofs-lens/internal"
+	common "github.com/nspcc-dev/neofs-node/cmd/neofs-lancet/internal"
 	"github.com/spf13/cobra"
 )
