+++ cmd/neofs-lancet/internal/meta/last_resync_epoch.go
@@ -1,7 +1,7 @@
 package meta
 
 import (
-	common "github.com/nspcc-dev/neofs-node/cmd/neofs-lens/internal"
+	common "github.com/nspcc-dev/neofs-node/cmd/neofs-lancet/internal"
 	"github.com/spf13/cobra"
 )
