+++ cmd/neofs-lancet/internal/object/link.go
