+++ cmd/neofs-lancet/internal/object/root.go
