+++ cmd/neofs-lancet/internal/printers.go
