+++ cmd/neofs-lancet/internal/storage/root.go
