+++ cmd/neofs-lancet/internal/storage/sanity.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/mr-tron/base58"
 	"github.com/nspcc-dev/bbolt"
-	common "github.com/nspcc-dev/neofs-node/cmd/neofs-lens/internal"
+	common "github.com/nspcc-dev/neofs-node/cmd/neofs-lancet/internal"
 	"github.com/nspcc-dev/neofs-node/cmd/neofs-node/config"
 	engineconfig "github.com/nspcc-dev/neofs-node/cmd/neofs-node/config/engine"
 	shardconfig "github.com/nspcc-dev/neofs-node/cmd/neofs-node/config/engine/shard"
