+++ cmd/neofs-lancet/internal/storage/write_cache_flush.go
@@ -1,7 +1,7 @@
 package storage
 
 import (
-	common "github.com/nspcc-dev/neofs-node/cmd/neofs-lens/internal"
+	common "github.com/nspcc-dev/neofs-node/cmd/neofs-lancet/internal"
 	"github.com/spf13/cobra"
 )
