+++ cmd/neofs-lancet/root.go
@@ -4,19 +4,19 @@ import (
 	"os"
 
 	"github.com/nspcc-dev/neofs-node/cmd/internal/cmderr"
-	"github.com/nspcc-dev/neofs-node/cmd/neofs-lens/internal/fstree"
-	"github.com/nspcc-dev/neofs-node/cmd/neofs-lens/internal/meta"
-	"github.com/nspcc-dev/neofs-node/cmd/neofs-lens/internal/object"
-	"github.com/nspcc-dev/neofs-node/cmd/neofs-lens/internal/storage"
+	"github.com/nspcc-dev/neofs-node/cmd/neofs-lancet/internal/fstree"
+	"github.com/nspcc-dev/neofs-node/cmd/neofs-lancet/internal/meta"
+	"github.com/nspcc-dev/neofs-node/cmd/neofs-lancet/internal/object"
+	"github.com/nspcc-dev/neofs-node/cmd/neofs-lancet/internal/storage"
 	"github.com/nspcc-dev/neofs-node/misc"
 	"github.com/nspcc-dev/neofs-node/pkg/util/gendoc"
 	"github.com/spf13/cobra"
 )
 
 var command = &cobra.Command{
-	Use:           "neofs-lens",
-	Short:         "NeoFS Storage Engine Lens",
-	Long:          `NeoFS Storage Engine Lens provides tools to browse the contents of the NeoFS storage engine.`,
+	Use:           "neofs-lancet",
+	Short:         "NeoFS Storage Engine Lancet",
+	Long:          `NeoFS Storage Engine Lancet provides tools to browse and change the contents of the NeoFS storage engine.`,
 	RunE:          entryPoint,
 	SilenceUsage:  true,
 	SilenceErrors: true,
@@ -25,7 +25,7 @@ var command = &cobra.Command{
 func entryPoint(cmd *cobra.Command, _ []string) error {
 	printVersion, _ := cmd.Flags().GetBool("version")
 	if printVersion {
-		cmd.Print(misc.BuildInfo("NeoFS Lens"))
+		cmd.Print(misc.BuildInfo("NeoFS Lancet"))
 
 		return nil
 	}
