+++ debian/neofs-cli.install
@@ -1,3 +1,3 @@
 bin/neofs-adm   usr/bin
 bin/neofs-cli   usr/bin
-bin/neofs-lens  usr/bin
+bin/neofs-lancet  usr/bin
